Move the event bindings out of App.ts
Third slice of the App.ts decomposition, after toolErrorMessages.ts and
statusBars.ts, and the same shape as both: a band of module-level functions
below the renderApp closure, moved without being rewritten.

What moves is all the DOM event wiring - the direct control bindings, the
delegated click/keydown handlers on the root and on document, the section,
info and advanced disclosure toggles, the tool-card view switcher, the
history action buttons, the external-link opener, and the data-openlayer-*
attribute lookups they share. It groups cleanly because it is pure DOM wiring
over an already-rendered tree: no generation state, no Photoshop calls, no
renderApp closure variables. The only outside references are AppElements,
AppView and setDiagnostics, all of which import.

The band is 614 lines and every one of them is byte-identical to what it was
in App.ts; the only edit is an export keyword on the 18 declarations App.ts
still needs. The seven attribute-lookup helpers stay module-internal because
nothing outside the band called them. App.ts drops from 5,514 to 4,917 lines.

The renderApp closure is untouched. That remains the part of this file that
does not get carved up.

Implemented by Codex against a written brief; reviewed, verified byte-identical
and validated by Claude. No review fixes were needed, so this branch is one
commit rather than the usual two.
